tonduyutung-bot/requirements.txt
python-telegram-bot
requests
axju-jokes
urbandictionary
chatterbot
chatterbot-corpus
